Shuffled code around a bit to make more readable
