--- src/Input/Input.ts
@@ -1,61 +1,0 @@
-import * as vscode from 'vscode';
-
-import { PushDataPoint, GetDataPointFromGUID } from '../Data/Data';
-import { DataPoint } from '../Data/DataPoint';
-
-import { getLineNumberFromUser } from './UserInput/getLineNumberFromUser';
-import { getDetailFromUser } from './UserInput/getDetailFromUser';
-import { getDataPointFromUser, getLinkedDataPointFromUser } from './UserInput/getDataPointFromUser';
-
-import { generateDataPointId } from './generateDataPointId';
-import { getActiveFile } from './getActiveFile';
-
-/**
- * Asynchronous function that creates a new DataPoint and adds it to the global state.
- */
-export async function CreateDataPoint() {
-  let dataPoint = new DataPoint();
-
-  dataPoint.lineNumber = await getLineNumberFromUser();
-  dataPoint.detail = await getDetailFromUser();
-  dataPoint.id = await generateDataPointId();
-
-  let linkedDataPoint = await getLinkedDataPointFromUser();
-
-  if (linkedDataPoint !== undefined) {
-    dataPoint.linkedDataPoints.push(linkedDataPoint.id);
-  }
-
-  dataPoint.file = await getActiveFile();
-
-  PushDataPoint(dataPoint);
-  vscode.window.showInformationMessage('Added data point with ID: ' + dataPoint.id);
-}
-
-/**
- * Asynchronous function that adds an additional link to an existing DataPoint and updates
- * it on the global state.
- */
-export async function AddAdditionalLinkToDataPoint() {
-  let currentDataPoint = await getDataPointFromUser();
-
-  if (currentDataPoint !== undefined) {
-    let linkedDataPoints: DataPoint[] = currentDataPoint.linkedDataPoints.map(dataPointGuid => {
-      return GetDataPointFromGUID(dataPointGuid);
-    });
-
-    let newLinkedDataPoint = await getLinkedDataPointFromUser([
-      currentDataPoint,
-      ...linkedDataPoints
-    ]);
-
-    if (newLinkedDataPoint) {
-      currentDataPoint.linkedDataPoints.push(newLinkedDataPoint.id);
-    }
-
-    PushDataPoint(currentDataPoint);
-    vscode.window.showInformationMessage(
-      'Added addtional link to Data Point: ' + currentDataPoint.id
-    );
-  }
-}
