+++ src/Output/Diagram/diagram.js
