+++ src/Output/Diagram/diagram.scss
