+++ src/Output/GenerateDiagram.ts
